move LandingPage into components subfolder
